% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dissimilarity.R
\name{calc_dissimilarity}
\alias{calc_dissimilarity}
\title{Calculate Dissimilarity Value to Reference Proteome for Peptides}
\usage{
calc_dissimilarity(
  pep,
  db = "human",
  k_val = 4.86936,
  a_val = 32,
  fill = NA_real_,
  tmp_dir = file.path(tempdir(), "neopeptides"),
  clean_tmp = TRUE
)
}
\arguments{
\item{pep}{a vector of peptides.}

\item{db}{an available database, can be one of "human" and "mouse".
Or you can set this to specify the (fasta) database file
to be searched instead of using standard database.
A blast database will be created if it does not exist.}

\item{k_val}{numeric. Steepness of sigmoidal curve at k.
Default 4.86936, the value used in the analysis of Van Allen,
Snyder, Rizvi, Riaz, and Hellmann datasets. See reference.}

\item{a_val}{numeric. Optionally can be "mean" to use mean alignment
for nmers passed. Horizontal displacement of partition function.
Default is 32, based on max_SW of 75 million 8-15mers from the
five clinical datasets against human, if using max_SW, use 52.
This value may not be meaningful for murine alignment so use with care.
See reference.}

\item{fill}{a numeric value for filling default NA value when no
blast result.}

\item{tmp_dir}{path for storing temp files.}

\item{clean_tmp}{if \code{TRUE}, remove temp directory.}
}
\value{
Data table of dissimilarity values (to the non-mutated proteome).
\itemize{
\item peptide - input peptide
\item dissimilarity - dissimilarity value
}
}
\description{
Calculate Dissimilarity Value to Reference Proteome for Peptides
}
\examples{
\donttest{
calc_dissimilarity("AAAAAAAAA")
calc_dissimilarity("MRLVDRRWA")
calc_dissimilarity("VRLVDRRWA")
calc_dissimilarity("MTEYKLVVVGAGDVGKSALTI")
calc_dissimilarity("MTEYKLVVVGAGDVGKSALTI", db = "mouse")
calc_dissimilarity(c("MTEYKLVVVGAGDVGKSALTIQLIQNHFVDEYDP", "MTEYKLVVVG"))
}
}
\references{
Richman LP, Vonderheide RH, and Rech AJ.
"Neoantigen dissimilarity to the self-proteome predicts immunogenicity
and response to immune checkpoint blockade." Cell Systems 9, 375-382.E4, (2019).
}
