% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_geometric_mean.R
\name{calc_geometric_mean}
\alias{calc_geometric_mean}
\title{Calculate the Geometric Mean}
\usage{
calc_geometric_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{ a vector or data.frame}

\item{na.rm}{remove NA values before processing}
}
\value{
The geometric mean(s)
}
\description{
Calculate the Geometric Mean
}
\examples{

x <- 1:5
calc_geometric_mean(x)
}
