% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_harmonic_mean.R
\name{calc_harmonic_mean}
\alias{calc_harmonic_mean}
\title{Calculate the Harmonic Mean}
\usage{
calc_harmonic_mean(x, na.rm = TRUE, zero = TRUE)
}
\arguments{
\item{x}{ a vector, matrix, or data.frame }

\item{na.rm}{na.rm=TRUE remove NA values before processing}

\item{zero}{If TRUE, then if there are any zeros, return 0, else, return the harmonic mean of the non-zero elements}
}
\value{
The harmonic mean(s)
}
\description{
Calculate the Harmonic Mean
}
\examples{

x <- 1:5
calc_harmonic_mean(x)
}
