% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_iedb_score.R
\name{calc_iedb_score}
\alias{calc_iedb_score}
\title{Calculate IEDB Score for Peptides}
\usage{
calc_iedb_score(
  pep,
  db = "human",
  fill = NA_real_,
  threads = parallel::detectCores(),
  tmp_dir = file.path(tempdir(), "neopeptides"),
  clean_tmp = TRUE
)
}
\arguments{
\item{pep}{a vector of peptides.}

\item{db}{an available database, can be one of "human" and "mouse".
Or you can set this to specify the (fasta) database file
to be searched instead of using standard database.
A blast database will be created if it does not exist.}

\item{fill}{a numeric value for filling default NA value when no
blast result.}

\item{threads}{number of threads to run.}

\item{tmp_dir}{path for storing temp files.}

\item{clean_tmp}{if \code{TRUE}, remove temp directory.}
}
\value{
Data table of IEDB scores.
\itemize{
\item peptide - input peptide
\item iedb_score - IEDB score
\item annotation - IEDB annotation info
}
}
\description{
Calculate IEDB Score for Peptides
}
\examples{
\donttest{
calc_iedb_score("AAAAAAAAA")
calc_iedb_score("MTEYKLVVVGAGDVGKSALTIQLIQNHFVDEYDP")
calc_iedb_score("MTEYKLVVVGAGDVGKSALTIQLIQNHFVDEYDP", db = "mouse")
calc_iedb_score(c("MTEYKLVVVGAGDVGKSALTIQLIQNHFVDEYDP", "MTEYKLVVVG"))
}
}
