% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_path}
\alias{find_path}
\title{Find Entry Path from Environment Variable or Config file}
\usage{
find_path(entry)
}
\arguments{
\item{entry}{a program or blast database.}
}
\value{
a path.
}
\description{
Firstly environment variable PATH will be searched, if not
found, configure file then will be used to search the entry.
}
\seealso{
\link{set_blast_path} for setting blast path, \link{install_database}
for installing database.
}
