% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neopeptides-package.R
\docType{package}
\name{neopeptides-package}
\alias{neopeptides}
\alias{neopeptides-package}
\title{neopeptides: Calculate and Explore Property Indices of Neopeptides}
\description{
Includes functions to calculate and explore several property indices for neopeptides, which are abnormal peptides generated from genome.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ShixiangWang/neopeptides}
  \item Report bugs at \url{https://github.com/ShixiangWang/neopeptides/issues}
}

}
\author{
\strong{Maintainer}: Shixiang Wang \email{w_shixiang@163.com} (\href{https://orcid.org/0000-0001-9855-7357}{ORCID})

}
\keyword{internal}
