% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_blast_path.R
\name{set_blast_path}
\alias{set_blast_path}
\title{Set Path to Blast}
\usage{
set_blast_path(path, store = TRUE)
}
\arguments{
\item{path}{a directory containing executable Blast program.
Installing it by conda is recommended.}

\item{store}{if \code{TRUE} (default), save the path to
config file so that you don't need to set this again when
start a new R session.}
}
\value{
Nothing
}
\description{
Set Path to Blast
}
\examples{
\donttest{
set_blast_path("/Users/wsx/anaconda3/bin/")
}
}
