% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_iedb_score.R
\name{calc_iedb_score}
\alias{calc_iedb_score}
\title{Calculate IEDB Score for Peptides}
\usage{
calc_iedb_score(
  pep,
  db = c("human", "mouse"),
  db_file = NULL,
  tmp_dir = file.path(tempdir(), "neopeptides"),
  clean_tmp = TRUE
)
}
\arguments{
\item{pep}{a vector of peptides.}

\item{db}{an available database, can be one of "human" and "mouse".}

\item{db_file}{you can set this to specify the (fasta) database file
to be searched instead of using standard database.}

\item{tmp_dir}{path for storing temp files.}

\item{clean_tmp}{if \code{TRUE}, remove temp directory.}
}
\value{
Data table of IEDB score
}
\description{
Calculate IEDB Score for Peptides
}
\examples{
\donttest{
calc_iedb_score("MTEYKLVVVGAGDVGKSALTIQLIQNHFVDEYDP")
calc_iedb_score("MTEYKLVVVGAGDVGKSALTIQLIQNHFVDEYDP", db = "mouse")
}
}
