% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_database.R
\name{install_database}
\alias{install_database}
\title{Install IEDB and Proteome Reference Database}
\usage{
install_database(
  db_path = "~/.neopeptide/db",
  force = FALSE,
  species = c("human", "mouse"),
  data_type = c("IEDB", "Proteome")
)
}
\arguments{
\item{db_path}{a path to store the databases.}

\item{force}{if \code{TRUE}, force to re-download and re-install.}

\item{species}{can be 'human' and 'mouse'. Default select them all.}

\item{data_type}{can be 'IEDB' and 'Proteome'. Default select them all.}
}
\value{
Nothing.
}
\description{
At default, configure file is '~/.neopeptide/config.yml', you
can set by \code{options(neopeptides.config="new_path.yml")}.
Please remember, if you change the configure file, you have
to set it per R session.
}
\examples{
\dontrun{
# Default all databases will be downloaded and installed.
install_database()
}
}
\seealso{
\link{set_blast_path}, \link{calc_iedb_score}, \link{calc_dissimilarity}
}
