% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_blast_path.R
\name{set_blast_path}
\alias{set_blast_path}
\title{Set Path to Blast}
\usage{
set_blast_path(path)
}
\arguments{
\item{path}{a directory containing executable Blast program.
Installing it by conda is recommended.}
}
\value{
Nothing
}
\description{
Set Path to Blast
}
\examples{
\dontrun{
set_blast_path("/Users/wsx/anaconda3/bin/")
}
}
\seealso{
\link{find_path}, \link{install_database}
}
